--- FoldingCell/FoldingCell/FoldingCell/Extensions.swift
@@ -1,54 +1,0 @@
-//
-//  Extensions.swift
-//
-// Copyright (c) 21/12/15. Ramotion Inc. (http://ramotion.com)
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// THE SOFTWARE.
-
-import UIKit
-
-public extension UIView {
-	
-	func pb_takeSnapshot(frame: CGRect) -> UIImage? {
-		
-		UIGraphicsBeginImageContextWithOptions(frame.size, false, 0.0)
-		
-		let context = UIGraphicsGetCurrentContext()
-		CGContextTranslateCTM(context, frame.origin.x * -1, frame.origin.y * -1)
-		
-		guard let currentContext = UIGraphicsGetCurrentContext() else {
-			
-			return nil
-		}
-		
-		layer.renderInContext(currentContext)
-		let image = UIGraphicsGetImageFromCurrentImageContext()
-		UIGraphicsEndImageContext()
-		
-		return image
-	}
-}
-
-internal extension UIColor {
-	
-	internal class func random() -> UIColor {
-		
-		return UIColor(red: CGFloat(arc4random_uniform(255))/255.0, green: CGFloat(arc4random_uniform(255))/255.0, blue: CGFloat(arc4random_uniform(255))/255.0, alpha: 1)
-	}
-}
